Moved server tests about
